fix uptime_sec overflowing after 4294 seconds m)

+++ webfrontend.cpp
@@ -29,7 +29,7 @@ int name2id(const char *fname, const int start = 0)
 
 uint32_t uptime_sec()
 {
-    return (uint32_t) esp_timer_get_time()/(int64_t)1000000;
+    return (esp_timer_get_time()/(int64_t)1000000);
 }
 
 String time_string(void)
